+++ OCSPatchers/Patchers/NewItems/LegendaryNPCItemsPatcher/EffectPatchers/Patchers/Clothing/DefencerSharpLegendaryItemEffect.cs
@@ -14,9 +14,9 @@ namespace OCSPatchers.Patchers.NewItems.LegendaryNPCItemsPatcher.EffectPatchers.
         public override bool TryApplyWeaponEffect(ModItem modItem)
         {
             if (!modItem.Values.ContainsKey(KEY_NAME)) return false;
-            if (modItem.Values[KEY_NAME] is not float originValue) return false;
+            if (modItem.Values[KEY_NAME] is not int originValue) return false;
 
-            modItem.Values[KEY_NAME] = (float)(originValue + 10);
+            modItem.Values[KEY_NAME] = (int)(originValue + 10);
 
             return true;
         }
